# rand: Use stdlib crypto/rand.Read on OpenBSD + Go 1.24