@jrick 2024-12-17
That would be a stupid change if they reverted it back, and not one I would anticipate, but it wouldn't break the correctness (it is still safe to concurrently call stdlib crypto/rand.Read)