@davecgh 2024-12-17
> That would be a stupid change if they reverted it back, and not one I would anticipate, but it wouldn't break the correctness (it is still safe to concurrently call stdlib crypto/rand.Read)

Agreed.  It's an unlikely change and correctness is the important part, particularly that `crypto/rand.Read` is now guaranteed not to error.  That why I said I don't really see any issue with the change.  I just wanted to point out the semantics change to being more dependent on the undocumented aspects in regards to performance.